chat son/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.5.0
pyodbc==4.0.39
flask-login==0.6.2
werkzeug==2.3.7
python-jose==3.3.0
passlib==1.7.4
faiss-cpu==1.7.4
markdown==3.5
langchain_community==0.0.10
faiss-cpu==1.7.4